% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_helpers.R
\name{arrange_on_page}
\alias{arrange_on_page}
\title{Arrange plots on a page}
\usage{
arrange_on_page(items = NULL, byrow = TRUE)
}
\arguments{
\item{items}{A list of items to arrange.}

\item{byrow}{Whether to fill rows or columns first per page.}
}
\value{
An integer vector of indices that can be used to sort \code{items} so that they
are neatly placed on a single page or multiple pages with \code{nrow} x \code{ncol}
places. The index is filled with \code{NA} on the last page.
}
\description{
This helper can be used to arrange plots with \code{\link[graphics:layout]{layout}}
on multiple pages. Unused positions on the grid are indexed with \code{NA}.
}
\details{
As a side-effect of calling this function, the \code{\link[graphics:layout]{layout}}
of the current device is set to a matrix of \code{nrow} x \code{ncol}.
}
\examples{
op <- graphics::par(no.readonly = TRUE)  # reconstitute par(...) afterwards
par(mfcol = c(4, 2), mar = rep(0, 4))
graphics::layout.show(9)  # one single plot on page 2

plot_me <- function(n) {

  for (i in n) {
    plot.new()
    text(0.5, 0.5, i)
  }

}

plot_me(seq(1, 3))  # appended on same page

par(mfcol = c(4, 2), mar = rep(0, 4))  # starts new page, but not always "accessible"

plot_me(arrange_on_page(seq(1, 9), byrow = FALSE))
plot_me(LETTERS[arrange_on_page(seq(1, 2), byrow = TRUE)])

rm(plot_me)
par(op)

}
