% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_helpers.R
\name{as_well}
\alias{as_well}
\title{Interpret characters as well names}
\usage{
as_well(well, as.tibble = FALSE, to.upper = TRUE, zero.padding = 2)
}
\arguments{
\item{well}{Character vector.}

\item{as.tibble}{If \code{TRUE}, a \code{\link[dplyr:tibble]{tibble}} is
returned instead of a list.}

\item{to.upper}{Should well letters be converted to uppercase if they are not?}

\item{zero.padding}{Number of digits the well number should be padded to.}
}
\value{
A list with \code{well_let}, \code{well_num} and \code{well} or a
tibble with the corresponding column names.
}
\description{
Interpret characters as well names
}
\examples{
as_well(c("a1", "B2", "Q23"))
as_well("23a")
as_well("n_2")

}
