% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_data.R
\docType{data}
\name{BobCAT}
\alias{BobCAT}
\title{Mass spectra for BobCAT treatment}
\format{
A list containing 12 \link[MALDIquant:MassSpectrum-class]{MassSpectrum}
objects.
}
\source{
Unpublished dataset.
}
\usage{
BobCAT
}
\description{
This dataset contains 12 \link[MALDIquant:MassSpectrum-class]{MassSpectrum}
objects collected upon treatment with a dilution series of BobCAT. The
duplicates are along wells with letter E and F respectively.
It is used to demonstrated the usage of \link{summerrmass-package}
}
\keyword{datasets}
