% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_helpers.R
\name{display_plate_layout}
\alias{display_plate_layout}
\title{Display a layout with ggplot2}
\usage{
display_plate_layout(layout, ...)
}
\arguments{
\item{layout}{A plate layout.}

\item{...}{Arguments passed to the \code{\link[ggplot2:geom_tile]{ggplot2::geom_tile()}}
aesthetic.}
}
\description{
The function returns the basic layer to be modified with further geoms by
the user, depending on the purpose.
}
\examples{
library(ggplot2)

layout_file <- system.file("extdata", "platelayout_maldi.xlsx", package = "summerrmass")

plate_384 <- import_layout_from_excel(layout_file, meta_row = c(Concentration = 1))
display_plate_layout(plate_384)  # all defined wells
display_plate_layout(plate_384, fill = content)
display_plate_layout(plate_384, fill = as.numeric(Concentration))
display_plate_layout(plate_384, fill = content, alpha = as.numeric(Concentration)) +
  labs(alpha = "concentration")

}
