% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_IC50.R
\name{fit_IC50}
\alias{fit_IC50}
\title{Fitting IC50 curves}
\usage{
fit_IC50(x, formula, newdata = NULL, ...)
}
\arguments{
\item{x}{A data frame to be evaluated.}

\item{formula}{A formula using variables of \code{x}.}

\item{newdata}{Data used for extrapolating of the generated model.}

\item{...}{Other arguments passed to \link{build_IC50}.}
}
\value{
A named list containing the tidied and augmented data frames.
}
\description{
Convenience wrapper calling \code{\link[broom:tidy]{broom::tidy}} and
\code{\link[broom:augment]{broom::augment}} upon the model specified in
\link{build_IC50}.
}
\details{
The IC50 curves are fitted for each group if \code{x} is a
\link[dplyr:grouped_df]{grouped data frame}.

Error and warning messages are mostly suppressed.
}
