% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_helpers.R
\name{get_border_indices}
\alias{get_border_indices}
\title{Get the indices of plots that are on page borders}
\usage{
get_border_indices(
  items = NULL,
  border = "b",
  byrow = TRUE,
  include.widows = TRUE
)
}
\arguments{
\item{items}{A list of items to arrange.}

\item{border}{Top (\code{"t"}), bottom (\code{"b"}), left (\code{"l"}), right
(\code{"r"}).}

\item{byrow}{Whether to fill rows or columns first per page.}

\item{include.widows}{\code{TRUE} if the last items on a semi-full page should
also be returned.}
}
\description{
Returns indices where \code{items} is next (in the direction of \code{border})
to a cell containing \code{NA}; all outer borders are also \code{NA}.
}
