% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maldi_fun.R
\name{get_layout_from_tree.maldi}
\alias{get_layout_from_tree.maldi}
\title{Establish the experiment layout from nested directories for MALDI analyses.}
\usage{
get_layout_from_tree.maldi(file_list, well_folder = "[0-9]_(?=[A-Z]+[0-9]+)")
}
\arguments{
\item{file_list}{A list of file paths or a character vector. See Details.}

\item{well_folder}{A \link[base:regex]{regular expression} describing the
names of the folders in a tree that contain measurements from a single well.
The well specifier must be excluded using a lookahead or lookbehind.}
}
\value{
A tibble with the experiment layout as determined from the paths in \code{file_list}
with columns \code{grp_0}, ..., \code{grp_N} specifiying the sample groups,
\code{well}, \code{replicate}, \code{n_replicates}, \code{path}.
}
\description{
This wrapper conveniently establishes the experiment layout from nested file
groups as exported by Bruker FlexAnalysis. Here, measurements for each plate
position (well) are saved in a folder such as "0_A1", "0_A2", etc. Replicate
measurements of the same well are grouped in subfolders such as "0_A1/1",
"0_A1/2" etc.
}
\details{
\code{file_path} can be a vector such as \code{c("file1", "file2", "file3")}
or a list of character vectors, e.g. after averaging spectra, such as
\code{list(c("file1", "file2"), "file3")}.

\code{well_folder} must be a unique match for each path.
}
\section{Sample groups by folder nesting}{


Sample groups are determined from the enclosing folders. For example, in

\preformatted{
/common1/folder1/folderA/0_A1/...
                      ../0_A2/...
              ../folderB/0_A1/...
/common1/folder2/folderA/0_A1/...
              ../folderB/0_A1/...
                        /0_A2/...
}

two nested groupings are identified: (1) "folder1" and "folder2" as \code{grp_0},
and (2) "folderA" and "folderB" as \code{grp_1}. The common folder "common1"
will not be used as a grouping variable. The first enclosing folder is always
used. The full (unique) grouping is given as \code{group}.

Grouping allows to anaylze multiple directories simultaneously (with their own or
a shared plate layout, see elsewhere).
}

