% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maldi_fun.R
\name{import_layout_from_paths.maldi}
\alias{import_layout_from_paths.maldi}
\title{Establish the experiment layout from nested directories for MALDI analyses}
\usage{
import_layout_from_paths.maldi(
  paths,
  pivot = "[0-9]_[A-Z]+[0-9]+",
  relative_to = getwd()
)
}
\arguments{
\item{paths}{A list character vector or list of file paths. See Details.}

\item{pivot}{A \link[base:regex]{regular expression} describing the
name of a single folder in each tree up to which "groups" and from which
"replicates" are established. See Details.}

\item{relative_to}{If not \code{NULL}, this subpath is hidden from the paths
for all operations; "" to suppress hiding; \code{NULL} will hide the subpath
that is common in all elements of \code{paths}.}
}
\description{
This wrapper conveniently establishes the experiment layout from nested file
groups as exported by Bruker FlexAnalysis. Here, measurements for each plate
position (well) are saved in a folder such as "0_A1", "0_A2", etc. Replicate
measurements of the same well are grouped in subfolders such as "0_A1/1",
"0_A1/2" etc.
}
\details{
\code{pivot} must contain a regular expression to identify the well, i.e.,
contain [A-Z]+[0-9]+".

The resulting \link[dplyr:grouped_df]{grouped data frame} is grouped by \code{well}.
}
\seealso{
\code{\link{import_layout_from_paths}}
}
