% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_loggers.R
\name{log_done}
\alias{log_done}
\title{Logging of process completion}
\usage{
log_done()
}
\description{
This logger can be conveniently activated/deactivatated by setting \code{options(summerr.log = ...)}.
}
