% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_loggers.R
\name{log_line}
\alias{log_line}
\title{Logging a separation line}
\usage{
log_line(char = "=", line = getOption("width"))
}
\arguments{
\item{char}{the character to repeat}

\item{line}{the length of the line}
}
\description{
This logger can be conveniently activated/deactivatated by setting \code{options(summerr.log = ...)}.
}
