% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_loggers.R
\name{log_process}
\alias{log_process}
\title{Logging of processes}
\usage{
log_process(...)
}
\arguments{
\item{...}{message}
}
\description{
Sub-routines performed under the headline of the current task.
}
\details{
This logger can be conveniently activated/deactivatated by setting \code{options(summerr.log = ...)}.
}
