% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_loggers.R
\name{log_task}
\alias{log_task}
\title{Logging of tasks}
\usage{
log_task(...)
}
\arguments{
\item{...}{message}
}
\description{
This logger can be conveniently activated/deactivatated by setting \code{options(summerr.log = ...)}.
}
