% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maldi_fun.R
\name{maldi_align_by_well}
\alias{maldi_align_by_well}
\title{Align and average MALDI spectra over repeated same-well measurements}
\usage{
maldi_align_by_well(
  object,
  method_baseline = "SNIP",
  method_average = "mean",
  ...,
  final_trim_range = c(0, Inf)
)
}
\arguments{
\item{object}{A list of \code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}}.}

\item{method_baseline}{Method for background subtraction.
See \link[MALDIquant:removeBaseline]{MALDIquant::removeBaseline}.}

\item{method_average}{Method to average replicate measurements.
See \link[MALDIquant:averageMassSpectra]{MALDIquant::averageMassSpectra}.}

\item{...}{further arguments passed to \code{\link[MALDIquant:alignSpectra]{alignSpectra}}}

\item{final_trim_range}{A numeric vector from which the maximum and the minimum
value are taken to trim the spectrum to a relevant range.}
}
\description{
Align and average MALDI spectra over repeated same-well measurements
}
