% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maldi_fun.R
\name{maldi_average_by_well}
\alias{maldi_average_by_well}
\title{Align and average MALDI spectra by well}
\usage{
maldi_average_by_well(
  object,
  method_baseline = "SNIP",
  method_average = "mean",
  ...,
  final_trim_range = c(0, Inf),
  pivot = "[0-9]_[A-Z]+[0-9]+"
)
}
\arguments{
\item{object}{A (list of) \code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}}.}

\item{method_baseline}{Method for background subtraction.
See \link[MALDIquant:removeBaseline]{MALDIquant::removeBaseline}.}

\item{method_average}{Method to average replicate measurements.
See \link[MALDIquant:averageMassSpectra]{MALDIquant::averageMassSpectra}.}

\item{...}{Arguments passed to \code{\link[MALDIquant:alignSpectra]{alignSpectra}}.}

\item{final_trim_range}{A numeric vector from which the maximum and the minimum
value are taken to trim the spectrum to a relevant range.}

\item{pivot}{A \link[base:regex]{regular expression} describing the
name of a single folder in each tree up to which "groups" and from which
"replicates" are established. See Details.}
}
\description{
Applies \link{maldi_average_spectra} to each group of wells using \link{maldi_tapply}.
}
\details{
The wells are identified using \code{pivot} from the associated paths in the
object. See \link{import_layout_from_paths.maldi}.
}
