% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maldi_fun.R
\name{maldi_average_spectra}
\alias{maldi_average_spectra}
\title{Align and average MALDI spectra}
\usage{
maldi_average_spectra(
  object,
  method_baseline = "SNIP",
  method_average = "mean",
  ...,
  final_trim_range = c(0, Inf)
)
}
\arguments{
\item{object}{A (list of) \code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}}.}

\item{method_baseline}{Method for background subtraction.
See \link[MALDIquant:removeBaseline]{MALDIquant::removeBaseline}.}

\item{method_average}{Method to average replicate measurements.
See \link[MALDIquant:averageMassSpectra]{MALDIquant::averageMassSpectra}.}

\item{...}{Arguments passed to \code{\link[MALDIquant:alignSpectra]{alignSpectra}}.}

\item{final_trim_range}{A numeric vector from which the maximum and the minimum
value are taken to trim the spectrum to a relevant range.}
}
\value{
A single \code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}}.
}
\description{
Default method to align and average MALDI spectra.
}
\details{
The baseline of the spectra is removed using \code{method_baseline} first,
then the spectra aligned using the default parameters of
\code{\link[MALDIquant:alignSpectra]{alignSpectra}} unless other arguments
are passed via \code{...}. The aligned spectra are averaged and trimmed to
\code{final_trim_range}.

This method can be substituted/overwritten to adapt specific needs in a pipeline.
}
