% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maldi_interface.R
\name{maldi_batch}
\alias{maldi_batch}
\title{Batch MALDI peak analysis of a directory}
\usage{
maldi_batch(
  path = NULL,
  pivot = "[0-9]_[A-Z]+[0-9]+",
  layout_file = NULL,
  confirm_layout_file = TRUE,
  manual = FALSE,
  review = TRUE,
  FUN_spect = maldi_average_by_well,
  MoreArgs_spect = list(pivot = "[0-9]_[A-Z]+[0-9]+"),
  FUN_peaks = maldi_find_peaks_by_well,
  MoreArgs_peaks = list(pivot = "[0-9]_[A-Z]+[0-9]+"),
  FUN_draw = maldi_draw_peaks_by_well,
  MoreArgs_draw = list(),
  FUN_import_layout = import_layout_from_excel,
  MoreArgs_layout = list(),
  MoreArgs_device = list(width = 21.5/2.54, height = 30.5/2.54, paper = "a4"),
  stored_peaks.generic = "maldi_peaks.rds",
  stored_spect.generic = "maldi_specs.rds",
  stamp = "\%y\%m\%d"
)
}
\arguments{
\item{path}{Directory with mass spectra to analyze, store and import results;
may contain the plate layout(s).}

\item{pivot}{A \link[base:regex]{regular expression} describing the
name of a single folder in each tree up to which "groups" and from which
"replicates" are established. See Details.}

\item{layout_file, confirm_layout_file, FUN_import_layout}{Path to a file that
contains metadata associated with each well/group. \code{layout_FUN} is applied
to \code{layout_file} before it is joined with the measurement result. Must
share column names with \link{import_layout_from_paths.maldi}. See Details.}

\item{manual}{...}

\item{review}{...}

\item{FUN_spect, MoreArgs_spect}{The function to be applied for pre-processing
the mass spectra using \code{MoreArgs_spect}.}

\item{FUN_peaks, MoreArgs_peaks}{The function to be applied for peak detection
using \code{MoreArgs_peak}.}

\item{stamp}{A datestamp (default) and/or timestamp constructed according
to this specification (\link[base:strptime]{POSIXct}) is added to pre-existing
files in path.}
}
\value{
A list with one element for each group.
}
\description{
This function is to control the flow of actions when batch processing a
directory. It establishes the save and recover policies of previously
analysed data so that the same data is not reanalyzed twice if new data
is added to a directory.
}
\details{
The batch process is split into different stages starting from the selected
\code{path} as main directory:

\enumerate{
\item{import and pre-processing of the spectra,}
\item{automated (or manual) peak picking for each spectrum,
depending if \code{manual = FALSE} (or \code{TRUE}),}
\item{review of automated peak picking if \code{review = TRUE}.}
}

These tasks are performed by group and the results are stored in the folder
enclosing all data of the same group. This is the parent to the folders that
match \code{pivot} (the well folders). See \link{import_layout_from_paths.maldi}.

If such a folder contains already analyses saved as \code{stored_peaks.generic}
or \code{stored_spect.generic}, these analyses can be imported for the next stage;
if not, the original files are renamed with a timestamp and new files added.

\subsection{Adapting the pipeline}{

...

\code{FUN_spect}, \code{FUN_peaks} and \code{layout_import_FUN} is can be of
type character or closure.

In case \code{FUN_spect} or \code{FUN_peaks} use a function that must determine,
e.g. which spectra are repeated measurements from a single well, which is
typically accomplished in the default functions using \link{import_layout_from_paths.maldi},
the \code{pivot} regex may need to be updated in the corresponding \code{MoreArgs_...}
arguments as well. A warning is emitted if it is not.

}

\subsection{Joining with metadata from external plate layout}{

If \code{layout_file} is \code{NULL}, the results will not be joined with an
external file that holds associated metadata. (Skip this section.)

External plate layouts can be provided via \code{layout_file} either (1) as an
absolute path to a single file somwhere in the file system or (2) a file in
expressed relative to the selected \code{path}, e.g. "global_layout.xls".
The query is made via \code{\link[base:files]{file.exists}}.
In such cases, the file is used for all groups. In case (3) when such a file
does not exist, the file is looked for locally in each group directory, where
it must be present each time under the same name, e.g., as "layout.xls".
If (4) this is neither the case and \code{confirm_layout_file = TRUE}, you
will finally be prompted to select a layout file. You can force this behavior
by setting \code{layout_file = ""}.

If \code{confirm_layout_file = TRUE}, you will also be asked if the file
identified according to one of the rules (1-3) above should be taken or not.

To successfully merge your layout with the data, common column names must
exist between your layout file and the result of \link{import_layout_from_paths.maldi},
which extends the generic \link{import_layout_from_paths} by explicitly
extracting a well number as column "well".

If you choose a graphical approach to define the plate layout, consider using
\link{import_layout_from_excel}, which will convert the plate into a (long)
table with a column "well" and "content".

If you provide such a (long) table directly, provide \code{\link{readxl::read_excel}}
or another appropriate function as argument \code{layout_import_FUN}.

}
}
\examples{
maldi_batch(
  MoreArgs_spect = list(pivot = "[0-9]_[A-Z]+[0-9]+",
                        final_trim_range = c(2420, 2445)),
  MoreArgs_peaks = list(pivot = "[0-9]_[A-Z]+[0-9]+",
                        mass_list = c(mC = 2425, hmC = 2441, fC = 2439),
                        tolerance_assignment = 0.5,
                        SNR = 3),
  MoreArgs_layout = list(ncol = 2, nrow = 6))

}
