% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maldi_fun.R
\name{maldi_draw_spectrum}
\alias{maldi_draw_spectrum}
\title{Plot a mass spectrum}
\usage{
maldi_draw_spectrum(
  object = NULL,
  xaxt = "s",
  x = NULL,
  y = NULL,
  overlay_color = NA,
  SNR = NaN,
  col = "red"
)
}
\arguments{
\item{object}{A \code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}}.}

\item{xaxt}{A character which specifies the x axis type. See \link[graphics:par]{par()}.}

\item{x, y}{The coordinates of additional points to draw.}

\item{overlay_color}{Color for overlay at 50\% transparency.}

\item{SNR}{Level at which the signal-to-noise-ratio should be indicated.}

\item{col}{Color for points and lines.}
}
\description{
Draws a mass spectrum using the default method and adds additional points and
lines if required.
}
\details{
If \code{object} is not a \code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}},
an empty plot is drawn.
}
