% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maldi_fun.R
\name{maldi_get_paths}
\alias{maldi_get_paths}
\title{Get the file paths of a MALDI spectrum as vector.}
\usage{
maldi_get_paths(object)
}
\arguments{
\item{object}{A (list of) \code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}}.}
}
\description{
Get the file paths of a MALDI spectrum as vector.
}
