% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maldi_fun.R
\name{maldi_import_spectra}
\alias{maldi_import_spectra}
\title{Import all MALDI spectra from a directory.}
\usage{
maldi_import_spectra(path = getwd(), ...)
}
\arguments{
\item{path}{\code{character}, path to directory or file which should be read
in.}

\item{...}{
  Arguments passed on to \code{\link[MALDIquantForeign:importMzXml-functions]{MALDIquantForeign::importMzXml}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
A \code{list} of \code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}}
objects. The scanned path is preserved as \code{attr(., "dir")}.
}
\description{
Calls \link[MALDIquantForeign:MALDIquantForeign-package]{MALDIquantForeign} to scan
a directory for MALDI spectra, import them, and perform intial qualtiy control.
}
\details{
In case a version of Bruker's CompassXport is installed, the user will be asked
whether the original FID files should be (re-)exported to the corresponding mzXML.

The spectra are checked for emptiness and regularity.
}
