% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maldi_fun.R
\name{maldi_peaks_by_well}
\alias{maldi_peaks_by_well}
\title{Extract intensities for given m/z values}
\usage{
maldi_peaks_by_well(
  object,
  mass_list,
  tolerance_assignment = 0.5,
  method = "MAD",
  SNR = 3,
  ...,
  manual = FALSE,
  pivot = "[0-9]_[A-Z]+[0-9]+"
)
}
\arguments{
\item{object}{A (list of) \code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}}.}

\item{mass_list}{Named list specifying the m/z at which to try extracting the
intensity.}

\item{tolerance_assignment}{Maximum difference between the calculated m/z and
the mass at the detected peak m/z.}

\item{method}{a noise estimation function; see
    \code{\link[MALDIquant]{estimateNoise,MassSpectrum-method}}.
  }

\item{SNR}{single numeric value. \code{SNR} is an abbreviation for
    \emph{s}ignal-to-\emph{n}oise-\emph{r}atio. A local maximum has to
    be higher than \code{SNR*noise} to be recognize as peak.}

\item{...}{Further arguments passed to \code{\link[MALDIquant:detectPeaks]{MALDIquant::detectPeaks}}.}

\item{manual}{Logical whether all peaks must be picked manually (\code{TRUE})
or only in ambiguous cases (\code{FALSE}).}

\item{pivot}{A \link[base:regex]{regular expression} describing the
name of a single folder in each tree up to which "groups" and from which
"replicates" are established. See Details.}
}
\value{
A \code{\link[tibble:tibble]{tibble}} with columns "ion", "mass", "intensity",
"percent" and well identifiers "well" etc. that relate the measured intensies
to the \code{object}, e.g. using "findex" (position in list) and/or "gindex"
(group number in list).
}
\description{
Extract intensities for given m/z values
}
\details{
Peaks are pre-selected from a smoothed spectrum within the m/z range given by
\code{mass_list} and \code{tolerance_assignment}. Peaks are measured in the
original, unsmoothed spectrum using \code{\link[MALDIquant:detectPeaks]{MALDIquant::detectPeaks}}.
}
\examples{
my_masses <- c(ion_species_A = 2400, ion_species_B = 2500)

}
