% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maldi_fun.R
\name{maldi_tapply}
\alias{maldi_tapply}
\title{Apply a function to a group of MALDI spectra}
\usage{
maldi_tapply(object, FUN, INDEX = NULL, MoreArgs = list())
}
\arguments{
\item{object}{A (list of) \code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}}.}

\item{FUN}{a function (or name of a function) to be applied, or \code{NULL}.
    In the case of functions like \code{+}, \code{\%*\%}, etc.,
    the function name must be backquoted or quoted.  If \code{FUN} is
    \code{NULL}, tapply returns a vector which can be used to subscript
    the multi-way array \code{tapply} normally produces.}

\item{INDEX}{a \code{\link[base]{list}} of one or more \code{\link[base]{factor}}s,
    each of same length as \code{X}.  The elements are coerced to
    factors by \code{\link[base]{as.factor}}.}

\item{MoreArgs}{A list of parameters passed to \code{FUN}.}
}
\value{
Depends on \code{FUN}.
}
\description{
Apply a function to a group of MALDI spectra
}
\details{
\code{FUN} is called with \code{MoreArgs} along the mass spectrum list \code{object}.
Thereby, \code{FUN} is called upon groups of \code{object}, if a grouping can
be established; else, \code{FUN} is applied for each element of \code{object}
separately.

To establish the grouping, \code{INDEX} is used.

\itemize{
\item{If \code{INDEX} is a \code{\link[base:data.frame]{data.frame}} or a similar
object such as a \code{\link[tibble:tibble]{tibble}}, the number and order of
the rows must correspond to the number and order of the elements in \code{object}.
}
\item{If \code{INDEX} qualifies as list of \code{\link[base:factor]{factor}}s,
each of the same length as \code{object}, this is used as grouping.
}
\item{If \code{INDEX} is \code{NULL}, an attempt is made to extract the
experiment layout from the file paths stored in the spectra (see
\link{import_layout_from_paths.maldi}) and the function is applied by well,
i.e., by a portion of the non-unique file path that matches \code{"[A-Z]+[0-9]+"}}.
}
}
