% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summerrmass.R
\name{maldi_template}
\alias{maldi_template}
\title{Get a template for a MALDI script}
\usage{
maldi_template(version = "A01")
}
\arguments{
\item{version}{A template version identifier.}
}
\description{
Get a template for a MALDI script
}
