% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_helpers.R
\name{model_cleanly_groupwise}
\alias{model_cleanly_groupwise}
\title{Fit, tidy and augment a model}
\usage{
model_cleanly_groupwise(x, FUN, newdata = NULL, ...)
}
\arguments{
\item{x}{Grouped data to apply \code{expr} upon. \code{x} is passed to \code{expr}
as first argument.}

\item{FUN}{A name of a function that returns a model object or another R object
with model information.}

\item{newdata}{Data points at which to evaluate the model.}

\item{...}{Arguments to \code{expr}.}
}
\value{
A \link[dplyr:grouped_df]{grouped data frame} with additional columns, \code{data},
\code{model}, \code{tidy}, \code{augment_old}, and \code{augment_new}.
}
\description{
Fits a modelling call to each group of a grouped data frame, extracts the
fitted parameters using \code{\link[broom:tidy]{broom::tidy}} and interpolates
missing values using \code{\link[broom:tidy]{broom::augment}} on existing points
or the \code{newdata}.
}
\details{
The fitting is wrapped with \code{\link[purrr:possibly]{purrr::possibly}} and
will return empty objects instead of errors.
}
