% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_helpers.R
\name{normalizePath}
\alias{normalizePath}
\title{Express file paths in their canonical form -- Truly}
\usage{
normalizePath(path)
}
\arguments{
\item{path}{A character vector of file paths.}
}
\value{
A character vector.
}
\description{
Under Windows and UNIX, \code{\link[base:.Platform]{.Platform$file.sep}} both
may expand to "\code{/}", whereas some upstream applications may still use
"\code{\\}" to separate folders and files.
}
\details{
Calls \code{\link[base:normalizePath]{normalizePath}} with \code{.Platform$file.sep}.
In addition, trailing slashes are removed.
}
