% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_helpers.R
\name{select_directory}
\alias{select_directory}
\title{Interactively choose a directory}
\usage{
select_directory(
  path = getwd(),
  caption = "Select a directory",
  label = "Select"
)
}
\arguments{
\item{path}{A character vector of length 1 to start browsing; defaults to the
current working directory.}

\item{caption}{The window title.}

\item{label}{The label to use for the 'Accept' / 'OK' button.}
}
\value{
A character with the path of the chosen directory.
}
\description{
This is a wrapper to select a directory \code{path} to process. Depending on
the mode R is invoked from, choosing is interactive and may use the RStudio IDE.
}
