% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_helpers.R
\name{select_single_file}
\alias{select_single_file}
\title{Interactively choose a single file}
\usage{
select_single_file(
  path = getwd(),
  prefix = ".+",
  suffix = "*",
  pattern = NULL,
  filetype = "file",
  caption = "Select a",
  label = "Select"
)
}
\arguments{
\item{path}{A character vector of file paths; the default corresponds to the
working directory, \code{\link[base]{getwd}()}. Tilde expansion (see \code{
\link[base]{path.expand}}) is performed.}

\item{prefix}{A regular expression to match the filename (without extension).}

\item{suffix}{A character to match the file extension, e.g. "csv".}

\item{pattern}{A \link[base:regex]{regular expression}. Only file names which
match the regular expression will be returned.}

\item{filetype}{A human-readable file type designation. Along with \code{suffix}
will be used to construct \code{filter} in \code{\link[rstudioapi:selectFile]{rstudioapi::selectFile}}.}

\item{caption}{The window title.}

\item{label}{The label to use for the 'Accept' / 'OK' button.}
}
\value{
A character vector with a single element in case a single file matches \code{pattern}
or a single choice was made, else an empty character vector.

In a non-interactive session, returns an empty character vector and a warning
if multiple files were found.
}
\description{
This is a convenient wrapper to select one file in the directory \code{path}
from multiple files matching \code{pattern}.
}
\details{
In case multiple files are found, the user is prompted to choose one option
or none.

If \code{pattern} is not \code{NULL}, \code{prefix} and \code{suffix} will be
ignored.

\code{prefix} is ignored when the RStudio API is used.
}
