% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summerrmass.R
\docType{package}
\name{summerrmass-package}
\alias{summerrmass-package}
\title{summerrmass}
\description{
Import and analyze peaks from MALDI mass spectra.
}
\details{
summerrmass-package



This package makes heavily use of \link{MALDIquant}. It
combines the functionality to pre-assembled "pipelines" that can be readily
used to batch process (nested) direcotries containing data from measurments
of plates. Each directory can be associated with a user-specified plate layout.
The metadata can be used in further analysis steps.
}
\section{General Functions}{


(These may be moved to a general package.)

\subsection{File and Directory Operations}{

\link{select_directory}

\link{select_single_file}

\link{backup_file}

} % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % %

\subsection{Importing Experiment Layouts}{

\link{import_layout_from_excel}

\link{import_layout_from_paths}

\link{display_plate_layout}

} % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % %

\subsection{Formatting Helpers}{

\link{as_well}

} % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % %
}

\section{Routines for MALDI peak calling from Burker FlexAnalysis}{


\link{maldi_import_spectra}

\link{maldi_average_spectra} is extended by \link{maldi_average_by_well}

\link{maldi_find_peaks_by_well} and \link{maldi_draw_peaks_by_well}

\link{maldi_tapply} a general wrapper to extend/apply functions over groups
such as wells

\subsection{Convenience Wrappers}{

\link{maldi_get_paths}

}
}

\section{Interactive Workflows}{


\link{maldi_batch}
}

\author{
Benjamin Buchmuller \email{benjamin.buchmuller@tu-dortmund.de}
}
