% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summerrmass.R
\docType{package}
\name{summerrmass}
\alias{summerrmass}
\title{summerrmass}
\description{
Import and analyse peaks from MALDI spectra.
}
\details{
summerrmass
}
\section{General Functions}{


(These may be moved to a general package.)

\subsection{Interaction}{

\link{select_directory}

\link{select_single_file}

} % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % %

\subsection{Import Experiment Layouts}{

\link{import_layout_from_excel}

\link{import_layout_from_paths}

} % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % %

\subsection{Formatting Helpers}{

\link{as_well}

} % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % % %
}

\section{Routines for MALDI peak calling from Burker FlexAnalysis}{


tbd
}

\section{Interactive Workflows}{


tbd
}

\author{
Benjamin Buchmuller \email{benjamin.buchmuller@tu-dortmund.de}
}
