% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summerrmass.R
\name{use_template}
\alias{use_template}
\title{Use a template for a MALDI script}
\usage{
use_template(version = "A01")
}
\arguments{
\item{version}{A template version identifier.}
}
\description{
Use a template for a MALDI script
}
\details{
\describe{
\item{A01}{MALDI batch processing for 384-well assay; used in Palei (2021)}
}
}
