% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_helpers.R
\name{wrap_augment}
\alias{wrap_augment}
\title{Wrapper: Augment}
\usage{
wrap_augment(expr, ...)
}
\arguments{
\item{expr}{A quoted model object or other R object with information.}

\item{...}{
  Arguments passed on to \code{\link[generics:augment]{generics::augment}}
  \describe{
    \item{\code{x}}{Model object or other R object with information to append to
observations.}
  }}
}
\description{
Passes a quoted expression to \code{\link[broom:augment]{broom::augment}}.
Errors are caught and an empty \code{tibble} is returned.
}
\details{
Note that you must pre-built the formula before passing it as \code{expr}.
}
\examples{
\dontrun{
FUN <- my_builder()
wrap_augment(FUN, ...)
}

}
