% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_helpers.R
\name{wrap_tidy}
\alias{wrap_tidy}
\title{Wrapper: Tidy}
\usage{
wrap_tidy(expr, ...)
}
\arguments{
\item{expr}{A quoted model object or other R object with information.}

\item{...}{
  Arguments passed on to \code{\link[generics:tidy]{generics::tidy}}
  \describe{
    \item{\code{x}}{An object to be converted into a tidy \code{\link[tibble:tibble]{tibble::tibble()}}.}
  }}
}
\description{
Passes a quoted expression to \code{\link[broom:augment]{broom::tidy}}.
Errors are caught and an empty \code{tibble} is returned.
}
\details{
Note that you must pre-built the formula before passing it as \code{expr}.
}
\examples{
\dontrun{
FUN <- my_builder()
wrap_augment(FUN, ...)
}

}
