% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_helpers.R
\name{backup_file}
\alias{backup_file}
\title{Backup a file if necessary}
\usage{
backup_file(path, sep = "_", stamp = "\%y\%m\%d")
}
\arguments{
\item{path}{A file path.}

\item{sep}{A character to separate the existing file path and the stamp added.}

\item{stamp}{A datestamp (default) and/or timestamp constructed according
to this specification (\link[base:strptime]{POSIXct}) is added to pre-existing
files in path.}
}
\value{
\code{TRUE} if succesful, \code{FALSE} otherwise.
}
\description{
Backup a file if necessary
}
\details{
The stamp is extracted from an existing file's modification time.
For details \link[base:file.info]{file.mtime(...)}.
If a backup file with the same stamp exists, it is replaced In practical terms
this means that by using \code{"\%y\%m\%d"} as stamp, only the last backup per
day is preserved.
}
