% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maldi_fun.R
\name{maldi_average_by_well}
\alias{maldi_average_by_well}
\title{Align and average MALDI spectra by well}
\usage{
maldi_average_by_well(
  object,
  INDEX = NULL,
  method_baseline = "SNIP",
  method_average = "mean",
  ...,
  final_trim_range = c(0, Inf)
)
}
\arguments{
\item{object}{A (list of) \code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}}.}

\item{INDEX}{a \code{\link[base]{list}} of one or more \code{\link[base]{factor}}s,
    each of same length as \code{X}.  The elements are coerced to
    factors by \code{\link[base]{as.factor}}.}

\item{method_baseline}{Method for background subtraction.
See \link[MALDIquant:removeBaseline]{MALDIquant::removeBaseline}.}

\item{method_average}{Method to average replicate measurements.
See \link[MALDIquant:averageMassSpectra]{MALDIquant::averageMassSpectra}.}

\item{...}{Arguments passed to \code{\link[MALDIquant:alignSpectra]{alignSpectra}}.}

\item{final_trim_range}{A numeric vector from which the maximum and the minimum
value are taken to trim the spectrum to a relevant range.}
}
\description{
Applies \link{maldi_average_spectra} to each group of wells using \link{maldi_tapply}.
}
