% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maldi_interface.R
\name{maldi_batch}
\alias{maldi_batch}
\title{Batch MALDI peak analysis of a directory}
\usage{
maldi_batch(
  path = NULL,
  manual = FALSE,
  review = TRUE,
  stored_peaks.generic = "maldi_peaks.rds",
  stored_spect.generic = "maldi_specs.rds",
  pivot = "[0-9]_[A-Z]+[0-9]+",
  MoreArgs = list()
)
}
\arguments{
\item{path}{Directory with mass spectra to analyze, store and import results;
may contain the plate layout.}

\item{manual}{...}

\item{review}{...}
}
\value{
A list with one element for each group.
}
\description{
This function is to control the flow of actions when batch processing a
directory. It establishes the save and recover policies of previously
analysed data so that the same data is not reanalyzed twice if new data
is added to a directory.
}
\details{
The batch process is split into different stages starting from the selected
\code{path} as main directory:

\enumerate{
\item{import and pre-processing of the spectra},
\item{automated (or manual) peak picking for each spectrum,
depending if \code{manual = FALSE} (or \code{TRUE})},
\item{review of automated peak picking if \code{review = TRUE}}.
}

These tasks are performed by group and the results are stored in the folder
enclosing all data of the same group. This is the parent to the folders that
match \code{pivot} (the well folders). See \link{import_layout_from_paths.maldi}.

If such a folder contains already analyses saved as \code{stored_peaks.generic}
or \code{stored_spect.generic}, these analyses can be imported for the next stage;
if not, the original files are renamed with a timestamp and new files added.

If such a folder contains a plate layout (see \link{import_layout_from_plate}),
the data is combined.
}
