% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maldi_fun.R
\name{maldi_draw_peaks_by_well}
\alias{maldi_draw_peaks_by_well}
\title{Draw m/z peaks in spectra}
\usage{
maldi_draw_peaks_by_well(
  object,
  data_peaks,
  file,
  title = NULL,
  ncol = 2,
  nrow = 6,
  highlight_missing_peaks = TRUE
)
}
\arguments{
\item{object}{A (list of) \code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}}.}

\item{data_peaks}{A peak data table as returned by \link{maldi_find_peaks_by_well};
must be a subset of \code{object}.}

\item{ncol}{Number of plot columns to arrange per page.}

\item{nrow}{Number of plot rows to arrange per page.}
}
\description{
Draws spectra and places them on a pdf page for printing.
}
\details{
No checking is done whether \code{data_peaks} was indeed derived from \code{object}.
However, \code{data_peaks} must contain a column \code{findex} which points to
the index of corresponding spectrum in \code{object}.
}
\examples{
data("BobCAT")

my_masses <- c(ion_species_A = 2425, ion_species_B = 2441)

# trim the spectra during averaging to speed-up the peak detection; alternatively
# use MALDIquant::trim(...).
my_spectrum <- maldi_average_by_well(BobCAT, final_trim_range = c(2420, 2445))
# detect peaks at the masses
my_peaks <- maldi_find_peaks_by_well(my_spectrum, my_masses)

}
